feat: Implement Vigenère encryption and decryption functions
The code changes add the implementation of Vigenère encryption and decryption functions. These functions allow for encrypting and decrypting messages using the Vigenère cipher algorithm. The functions take a message and a key as input and return the encrypted or decrypted message, respectively.

The commit message follows the established convention of using a prefix to indicate the type of change. In this case, "feat" is used to indicate a new feature implementation. The message is concise and clearly describes the purpose of the code changes.